PWM Control

+++ embedded/Actuator/PicoPWM/pwm_control.cpp
@@ -26,7 +26,7 @@ extern "C" {
         // Tell the LED pin that the PWM is in charge of its value.
         gpio_set_function(GPIO_PIN, GPIO_FUNC_PWM);
         // Figure out which slice we just connected to the LED pin
-        uint slice_num = pwm_gpio_to_slice_num(GPIO_PIN);
+        uint8_t slice_num = pwm_gpio_to_slice_num(GPIO_PIN);
 
         // Mask our slice's IRQ output into the PWM block's single interrupt line,
         // and register our interrupt handler
